refactor(dom): Add <nav> to store list of projects
Add a sidebar nav and a primary nav for storing <ul> list of projects.
This helps to just append the list inside one of them instead of trying
to put it after a sibling or prepending it. This <nav> tag is created in
template.html and is not removed after that.

- nav container for projects list