refactor(html): Add containers for hamburger button
Add two containers one in main page and another in sidebar to contain
hamburger button. In the main page it doesn't acually contain it but
rather acts as an empty space to push the heading down so that it
doesn't cover hamburger button. In future some features can also be
added to it.

- hamburger container
- topheading container